pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.2",
    "eventlet>=0.40.4",
    "flask>=3.1.3",
    "flask-cors>=6.0.2",
    "flask-socketio>=5.6.1",
    "numpy>=2.4.2",
    "openalgo>=1.0.46",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
